Added database for new MLP version data
I downloaded the v2.0 linked dataset using the IPUMS extract system
instead of the API. The current scripts write the extracted data to a
new database file.

This should probably change to having a script to pull the necessary
variables, but I didn't know what I would need so I grabbed as much as I
could for now.